python-sdk/examples/llamaindex/requirements.txt
xproof>=0.1.0
llama-index-core>=0.10.0
